Provera1

+++ app/src/main/java/com/example/lotto/data/LottoOffer.kt
@@ -8,7 +8,7 @@ data class LottoOffer (
     @SerializedName("eventId"            ) var eventId            : Int?    = null,
     @SerializedName("gameId"             ) var gameId             : Int?    = null,
     @SerializedName("name"               ) var name               : String? = null,
-    @SerializedName("time"               ) var time               : Int?    = null,
+    @SerializedName("time"               ) var time               : Long?    = null,
     @SerializedName("allowBetTimeBefore" ) var allowBetTimeBefore : Int?    = null
 
 )
